requirements.txt
oracledb
langchain
langchain-community
langchain-core
python-dotenv
sqlalchemy
